♻️ move `PoliSearchBar` out of the `Home` folder
