+++ src/components/Home/index.tsx
@@ -2,5 +2,4 @@ export * from "./MainMenu"
 export * from "./MainTitle"
 export * from "./MenuButton"
 export * from "./StickyHeader"
-export * from "./PoliSearchBar"
 export * from "./SearchButton"
